src/streamlit-webapp/requirements.txt
Pillow
opencv-python
roboflow
ultralytics